third_party/blink/web_tests/svg/dynamic-updates/SVGFEDiffuseLightingElement-dom-in-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'in' attribute of the SVGFEDiffuseLightingElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS gradientElement.getAttribute('in') is "SourceGraphic"
PASS gradientElement.getAttribute('in') is "blur"
PASS successfullyParsed is true

TEST COMPLETE
